Add the copyleft-elkstack cookbook
The copyleft-elkstack cookbook is designed to install and configure a
basic elkstack for a 'Proof of Concept' environment. It uses a CentOS
7 base VM and install packages from Elastic's public yum repositories.

'kitchen converge' will create a virtual machine with ELK installed,
Elastic listening on  port 9200, LogStash listening for Beats formatted
messages on port 5044, and Kibana listening on port 5600.

Remove jdk artifact from copyleft-elkstack